changed query parameters types and added perceived objects trigger

+++ database/mssql/procedures/read/denm.sql
@@ -3,7 +3,9 @@ DROP PROCEDURE get_denms;
 DROP PROCEDURE get_denms_quadtree;
 DROP PROCEDURE get_denms_quadtree_and_station_id; 
 GO 
-CREATE PROCEDURE get_denms @time_start INT, @time_end INT
+CREATE PROCEDURE get_denms 
+  @time_start BIGINT, 
+  @time_end BIGINT
 AS
 select 
   DENM.event_timestamp as [timestamp],
@@ -18,7 +20,10 @@ from
 where 
   DENM.event_timestamp between @time_start and @time_end
 GO
-CREATE PROCEDURE get_denms_station_id @time_start INT, @time_end INT, @in_station_id INT
+CREATE PROCEDURE get_denms_station_id 
+  @time_start BIGINT, 
+  @time_end BIGINT, 
+  @in_station_id BIGINT
 AS
 select 
   DENM.event_timestamp as [timestamp],
@@ -34,7 +39,11 @@ where
   DENM.event_timestamp between @time_start and @time_end and
   @in_station_id = emitter_station_id
 GO
-CREATE PROCEDURE get_denms_quadtree @time_start INT, @time_end INT, @quadtree_start BIGINT, @quadtree_end BIGINT
+CREATE PROCEDURE get_denms_quadtree 
+  @time_start BIGINT, 
+  @time_end BIGINT, 
+  @quadtree_start BIGINT, 
+  @quadtree_end BIGINT
 AS
 select 
   DENM.event_timestamp as [timestamp],
@@ -50,7 +59,12 @@ where
   DENM.event_timestamp between @time_start and @time_end and
   quadtree between @quadtree_start and @quadtree_end
 GO
-CREATE PROCEDURE get_denms_quadtree_and_station_id @time_start INT, @time_end INT, @quadtree_start BIGINT, @quadtree_end BIGINT, @in_station_id INT
+CREATE PROCEDURE get_denms_quadtree_and_station_id 
+  @time_start BIGINT, 
+  @time_end BIGINT, 
+  @quadtree_start BIGINT, 
+  @quadtree_end BIGINT, 
+  @in_station_id BIGINT
 AS
 select 
   DENM.event_timestamp as [timestamp],
